Updated to the API json-1.0 branch.

+++ src/main/java/javax/json/stream/JsonLocation.java
@@ -23,14 +23,14 @@ public interface JsonLocation {
      *
      * @return the line number or -1 if none is available
      */
-    int getLineNumber();
+    long getLineNumber();
 
     /**
      * Return the column number for the current JSON event in the input source.
      *
      * @return the column number or -1 if none is available
      */
-    int getColumnNumber();
+    long getColumnNumber();
 
     /**
      * Return the stream offset into the input source this location
@@ -42,6 +42,6 @@ public interface JsonLocation {
      * @return the offset of input source stream, or -1 if there is
      * no offset available
      */
-    int getStreamOffset();
+    long getStreamOffset();
 
 }
